Refactor: 기존 UserService를 주입받는 대상을 변경된 UserServiceImpl이 아닌 UserService 인터페이스로 수정
